requirements.txt
configparser
gitpython
pyyaml
github3.py==1.0.0a4
